DemoAutomation struct for managing cloud client and getting events
