+++ .env.example
@@ -1,16 +1,7 @@
-DSN=
 DSN_PYTHON=
-DSN_REACT=
-DSN_ANDROID=
 
 DSN_PYTHON_SAAS=
 DSN_REACT_SAAS=
-DSN_EXPRESS_SAAS=
-DSN_RUBY_SAAS=
-
-DSN_PYTHON_GATEWAY=
-DSN_PYTHON_CELERY=
-DSN_PYTHON_DJANGO=
 
 SQLITE=
