+++ main.go
@@ -2,36 +2,38 @@ package main
 
 import (
 	"context"
-	"encoding/json"
+	// "encoding/json"
 	"flag"
-	"fmt"
-	"io/ioutil"
+
+	// "io/ioutil"
 	"log"
 	"net/http"
 	"os"
-	"time"
 
 	"cloud.google.com/go/storage"
 	"github.com/joho/godotenv"
 	_ "github.com/mattn/go-sqlite3"
-	"google.golang.org/api/iterator"
+	// "google.golang.org/api/iterator"
 )
 
-var httpClient = &http.Client{}
+var httpClient = &http.Client{} // TODO
 
 var (
-	all         *bool
-	id          *string
-	ignore      *bool
-	database    string
-	db          *string
-	js          *string
-	py          *string
-	dsn         DSN
+	all    *bool
+	id     *string
+	ignore *bool
+	db     *string // or similar could be for dynamically handling the 'prefix' for Bucket
+	js     *string
+	py     *string
+	// dsn         DSN
 	SENTRY_URL  string
-	exists      bool
 	projectDSNs map[string]*DSN
 	traceIds    []string
+
+	// TESTING...
+	client       *storage.Client
+	ctx          context.Context
+	bucketHandle *storage.BucketHandle
 )
 
 func init() {
@@ -42,12 +44,10 @@ func init() {
 	all = flag.Bool("all", false, "send all events. default is send latest event")
 	// id = flag.String("id", "", "id of event in sqlite database") // 08/27 non-functional today
 	ignore = flag.Bool("i", false, "ignore sending the event to Sentry.io")
-	// db = flag.String("db", "", "database.json")
 	js = flag.String("js", "", "javascript DSN")
 	py = flag.String("py", "", "python DSN")
 	flag.Parse()
 
-	// TODO demoAutomation.Dsns.configure() or demoAutomation.configureDsns()
 	projectDSNs = make(map[string]*DSN)
 	projectDSNs["javascript"] = parseDSN(os.Getenv("DSN_JAVASCRIPT_SAAS"))
 	if *js != "" {
@@ -57,84 +57,13 @@ func init() {
 	if *py != "" {
 		projectDSNs["python"] = parseDSN(*py)
 	}
-
-	// if *db == "" {
-	// 	database = os.Getenv("JSON")
-	// } else {
-	// 	database = *db
-	// }
 }
-
-type DemoAutomation struct {
-	client                *storage.Client
-	ctx                   context.Context
-	bucketHandle          *storage.BucketHandle // `client.Bucket(bucketName)` for setting this
-	bucketHandleFileNames []string              // `query := &storage.Query{Prefix: "eventtest"}` for setting this
-	// TODO consider `events []Event` ?
-	// TODO consider setDsns...
-}
-
-// TODO METHODS
-// Constructor/Init for DemoAutomation
-// then can call demoAutomation.client() demoAutomation.bucketHandle.Objects()
 
 func main() {
-	// TODO da | demoAutomation
 
-	bucketName := os.Getenv("BUCKET")
-	// Initialize/Connect the Client
-	ctx := context.Background()
-	client, err := storage.NewClient(ctx)
-	if err != nil {
-		log.Fatalln("storage.NewClient:", err)
-	}
-	defer client.Close()
-	ctx, cancel := context.WithTimeout(ctx, time.Second*50)
-	defer cancel()
-	// Prepare bucket handle
-	bucketHandle := client.Bucket(bucketName)
-	// lists the contents of a bucket in Google Cloud Storage.
-	var fileNames []string
-	query := &storage.Query{Prefix: "eventtest"}
-	it := bucketHandle.Objects(ctx, query)
-	for {
-		obj, err := it.Next()
-		if err == iterator.Done {
-			break
-		}
-		if err != nil {
-			log.Fatalln("listBucket: unable to list bucket", err)
-		}
-		fileNames = append(fileNames, obj.Name)
-		printObj(obj)
-	}
+	demoAutomation := DemoAutomation{}
 
-	// TODO events.go could manage reading from storage. or like:
-	/*
-		storageClient := StorageClient(os.Getenv("BUCKET")) <-- is the init
-		//or
-		demoAutomation.init(os.Getenv("BUCKET"))
-		demoAutomation.query("event") .prefixQuery("event") .queryBucket .bucketQuery() .bucketSet()
-		demoAutomation.listBucketContents() .getBucket()
-		events := demoAutomation.getFiles() .bucketFiles()
-	*/
-
-	// Read each file's content
-	var events []Event
-	for _, fileName := range fileNames {
-		rc, err := bucketHandle.Object(fileName).NewReader(ctx)
-		if err != nil {
-			log.Fatalln("NewReader:", err)
-		}
-		byteValue, _ := ioutil.ReadAll(rc)
-
-		// The Event's UnmarshalJSON overriden in Event.go
-		var event Event
-		if err := json.Unmarshal(byteValue, &event); err != nil {
-			panic(err)
-		}
-		events = append(events, event)
-	}
+	events := demoAutomation.getEvents() // configureDsns() too ?
 
 	for _, event := range events {
 		if event.Kind == "error" {
@@ -154,16 +83,63 @@ func main() {
 	getTraceIds(events)
 	updateTraceIds(events)
 
-	// TODO double check it's object was updated reference `fmt.Println("\n> timestamp AFTER", event.Error.Timestamp)`
 	requests := Requests{events}
 	requests.send()
 
 	return
 }
 
-func printObj(obj *storage.ObjectAttrs) {
-	fmt.Printf("filename: /%v/%v \n", obj.Bucket, obj.Name)
-	// fmt.Printf("ContentType: %q, ", obj.ContentType)
-	// fmt.Printf("Owner: %v, ", obj.Owner)
-	// fmt.Printf("Size: %v, ", obj.Size)
-}
+// demoAutomation.init()
+
+// // 1 ctx, client
+// ctx := context.Background()
+// client, err := storage.NewClient(ctx)
+// if err != nil {
+// 	log.Fatalln("storage.NewClient:", err)
+// }
+// defer client.Close()
+// ctx, cancel := context.WithTimeout(ctx, time.Second*50)
+// defer cancel()
+
+// // 2 bucket handle
+// bucketName := os.Getenv("BUCKET")
+// bucketHandle := client.Bucket(bucketName)
+
+// lists the contents of a bucket in Google Cloud Storage.
+// demoAutomation.getFileNames()
+
+// var fileNames []string
+// query := &storage.Query{Prefix: "eventtest"}
+// it := demoAutomation.BucketHandle.Objects(demoAutomation.Ctx, query)
+// // it := bucketHandle.Objects(ctx, query)
+// for {
+// 	obj, err := it.Next()
+// 	if err == iterator.Done {
+// 		break
+// 	}
+// 	if err != nil {
+// 		log.Fatalln("listBucket: unable to list bucket", err)
+// 	}
+// 	fileNames = append(fileNames, obj.Name)
+// 	printObj(obj)
+// }
+
+// Read each file's content
+// var events []Event
+// for _, fileName := range fileNames {
+// 	rc, err := demoAutomation.BucketHandle.Object(fileName).NewReader(demoAutomation.Ctx)
+// 	// rc, err := bucketHandle.Object(fileName).NewReader(ctx)
+// 	if err != nil {
+// 		log.Fatalln("NewReader:", err)
+// 	}
+// 	byteValue, _ := ioutil.ReadAll(rc)
+
+// 	// The Event's UnmarshalJSON overriden in Event.go
+// 	var event Event
+// 	if err := json.Unmarshal(byteValue, &event); err != nil {
+// 		panic(err)
+// 	}
+// 	events = append(events, event)
+// }
+
+// TODO double check it's object was updated reference `fmt.Println("\n> timestamp AFTER", event.Error.Timestamp)`
